build: land cl_pose_mp/cl_net_chan_mp/sv_main_pc_mp/g_scr_mover
POSIX build now ~92 upstream files. Same files on Switch.

client_mp/cl_pose_mp.cpp: clean once InterlockedExchangeAdd shim was
added to kisak_compat.h (atomic-fetch-add via __atomic_fetch_add).

client_mp/cl_net_chan_mp.cpp:
- 2 [[maybe_unused]] on hex-rays scratch locals (iTotalBPS, iYStep).
- if (clientConnections) → if (true). clientConnections is a fixed-
  size array; address is always truthy. Upstream's check predates the
  static array layout.

server_mp/sv_main_pc_mp.cpp: 2 KISAKHACK casts on
com_masterServerName->current.integer to bridge 32→64 bit pointer.

game/g_scr_mover.cpp: 1 precedence-paren fix.

qcommon/com_bsp_load_obj.cpp: initialized 'zero' variable that was
previously uninitialized — clang's newer -Wuninitialized-const-pointer
caught it after FS_Write's signature changes earlier in this session.

posix/kisak_compat.h: InterlockedExchangeAdd template via
__atomic_fetch_add (seq-cst), returns previous value.

posix_backbone_stubs.cpp additions (~20 entries):
- AxisToAngles, CL_DrawString, G_DObjUpdate, G_FreeEntity,
  I_dmaGetDObjSkel.
- NET_AdrToString, NET_CompareBaseAdr, NET_OutOfBandPrint (bool sig),
  NET_StringToAdr.
- Scr_GetNumParam, Scr_GetVector, Scr_Notify, Scr_ObjectError,
  Scr_ParamError.
- SV_FindClientByAddress, SV_PreGameUserVoice, SV_UserVoice,
  SVC_GameCompleteStatus.
- cl_profileTextHeight, rcon_password (dvar* globals).

Deliberately deferred (compile-clean but drag heavy stub families,
ready for bigger batches later):
- client_mp/cl_input + cl_parse_mp (20+ cl_* dvars, MSG_Write*).
- server_mp/sv_snapshot_mp (MSG_WriteEntity/Delta*, SV_*Track*).
- game/g_helicopter + g_hudelem + g_items (vehHelicopter dvars,
  BG_* helpers from bg_pmove/bg_weapons).
- cgame/cg_ammocounter + cg_effects_load_obj + cg_laser + cg_shellshock
  (bgame dvars + Hunk_AllocateTempMemoryHigh + BG_AmmoForWeapon family).

+++ src/server_mp/sv_main_pc_mp.cpp
@@ -17,9 +17,11 @@ const netadr_t *__cdecl SV_MasterAddress()
     if (adr.type == NA_BOT)
     {
         Com_Printf(15, "Resolving %s\n", com_masterServerName->current.string);
-        if (NET_StringToAdr((char *)com_masterServerName->current.integer, &adr))
+        // KISAKHACK: DvarValue.integer holds 32-bit pointer in upstream;
+        // bridge through uintptr_t.
+        if (NET_StringToAdr((char *)(uintptr_t)com_masterServerName->current.integer, &adr))
         {
-            const char* result = strstr(":", (char*)com_masterServerName->current.integer);
+            const char* result = strstr(":", (char*)(uintptr_t)com_masterServerName->current.integer);
             if (!result)
                 adr.port = BigShort(com_masterPort->current.integer);
             v1 = BigShort(adr.port);
